Extract Benefits into component
