+++ pages/careers/senior-product-designer.js
@@ -1,56 +1,20 @@
-import styled, { css } from 'styled-components'
+import styled from 'styled-components'
 import Br from 'components/Br'
 
 import Page, { page } from 'components/Page'
+import { Section, SectionHeader, Subsection, SubSectionHeader } from 'components/sections/sections'
 import BodyCopy from 'components/BodyCopy'
 import { Masonry, Brick } from 'components/Masonry'
 import { LinkList, LinkListItem, LinkListItemLink } from 'components/LinkList'
 import Callout from 'components/Callout'
+import Benefits, { CareersParagraph } from 'components/Benefits'
 
 import headOfficeImage from '../../assets/images/about/head-office.jpg'
-
-import healthCoverIcon from '../../assets/images/careers/streamline-icon-hospital-house@48x48.svg'
-import homeOfficeIcon from '../../assets/images/careers/streamline-icon-office-desk@48x48.svg'
-import learningIcon from '../../assets/images/careers/streamline-icon-tools-box@48x48.svg'
-import parentalLeaveIcon from '../../assets/images/careers/streamline-icon-baby-care-body@48x48.svg'
-import profitShareIcon from '../../assets/images/careers/streamline-icon-saving-piggy-dollars@48x48-2.svg'
-import wellnessIcon from '../../assets/images/careers/streamline-icon-messages-bubble-smile@48x48.svg'
-
-const Section = styled.section`
-  margin-top: ${({ theme }) => theme.outerSpacing.s3};
-  margin-bottom: ${({ theme }) => theme.outerSpacing.s3};
-
-  ${(props) => props.small && css`
-    margin: ${({ theme }) => theme.outerSpacing.s3} auto;
-    max-width: 510px;
-  `}
-`
-
-const SectionHeader = styled.h2`
-  ${({ theme }) => theme.textStyles.secondLevelHeading}
-  margin-top: ${({ theme }) => theme.innerSpacing.s2};
-  text-align: center;
-`
-
-const Subsection = styled.section`
-  margin-top: ${({ theme }) => theme.innerSpacing.s2};
-  margin-bottom: ${({ theme }) => theme.innerSpacing.s2};
-`
-
-const SubSectionHeader = styled.h3`
-  ${({ theme }) => theme.textStyles.bodyCopy}
-  margin: ${({ theme }) => theme.textSpacing.s2} 0 ${({ theme }) => theme.textSpacing.s1} 0;
-`
 
 const CareersSectionParagraph = styled.p`
   ${({ theme }) => theme.textStyles.bodyCopy};
   color: ${({ theme }) => theme.colors.text.subdued};
   margin: ${({ theme }) => theme.textSpacing.s3} 0 0;
-`
-
-const CareersParagraph = styled.p`
-  color: ${({ theme }) => theme.colors.text.subdued};
-  margin: ${({ theme }) => theme.textSpacing.s1} 0;
 `
 
 const ApplicationProcess = styled.dl`
@@ -117,15 +81,15 @@ export default page((props) => (
           <tr>
             <th>Salary:</th>
             <td>
-              AUD $130K &mdash; $160K salary<br/>
-              USD $90K &mdash; $110K salary<br/>
-              €80 &mdash; 100K salary<br/>
+              AUD $130K &mdash; $160K salary<br />
+              USD $90K &mdash; $110K salary<br />
+              €80 &mdash; 100K salary<br />
               Depending on experience, plus benefits, for 38 hours per week (pro-rata for part-time). Salary amount excludes superannuation / 401k.
             </td>
           </tr>
           <tr>
             <th>Location:</th>
-            <td>Remote with at least 3 hours of overlap with UTC+10 mornings (you'll be working with people in Vancouver, Perth, and Melbourne)</td>
+            <td>Remote with at least 3 hours of overlap with UTC+10 mornings (you&rsquo;ll be working with people in Vancouver, Perth, and Melbourne)</td>
           </tr>
           <tr>
             <th>Hours:</th>
@@ -141,7 +105,7 @@ export default page((props) => (
 
         <p>We are an independently owned, profitable company that values sustainable growth, transparency and work-life balance. We&lsquo;re currently a distributed team of eleven, and we work remotely from Melbourne, Vancouver, Perth, Sydney, and Adelaide.</p>
 
-        <p>As a Senior Product Designer at Buildkite you'll work on a small team with a big impact, and help shape a tool that thousands of developers use every day. You'll report to Tim Lucas, our co-founder and design lead, and you'll be working alongside an existing senior product designer and print designer. In this small team you'll have the chance to influence all corners of Buildkite's product and brand.</p>
+        <p>As a Senior Product Designer at Buildkite you&rsquo;ll work on a small team with a big impact, and help shape a tool that thousands of developers use every day. You&rsquo;ll report to Tim Lucas, our co-founder and design lead, and you&rsquo;ll be working alongside an existing senior product designer and print designer. In this small team you&rsquo;ll have the chance to influence all corners of Buildkite&rsquo;s product and brand.</p>
 
         <h3>A typical day for a Senior Product Designer might look like:</h3>
 
@@ -153,8 +117,8 @@ export default page((props) => (
           <li>Writing a post to the team on Basecamp with some design strategies and ideas, and getting feedback from a range of perspectives on the team.</li>
           <li>Giving feedback on and helping to iterate our UI component libraries, helping to empower the rest of the team with great design tools.</li>
           <li>Creating polished design assets for product launches, and designing changes to the Buildkite marketing site.</li>
-          <li>oining in the weekly design catchup over Zoom, giving feedback on a commissioned illustration project for a new conference sticker, contributing to a sponsorship booth design idea, and demonstrating designs you've been working on.</li>
-          <li>Pairing on support to answer customer questions, and venturing into the far corners of Buildkite's product, tools and documentation to understand how things are working in practice.</li>
+          <li>Joining in the weekly design catchup over Zoom, giving feedback on a commissioned illustration project for a new conference sticker, contributing to a sponsorship booth design idea, and demonstrating designs you&rsquo;ve been working on.</li>
+          <li>Pairing on support to answer customer questions, and venturing into the far corners of Buildkite&rsquo;s product, tools and documentation to understand how things are working in practice.</li>
         </ul>
 
         <h3>This job is for you if you:</h3>
@@ -207,44 +171,7 @@ export default page((props) => (
     <Section>
       <SectionHeader>Benefits</SectionHeader>
       <Subsection>
-        <Masonry>
-          <Brick
-            icon={learningIcon}
-          >
-            <SubSectionHeader>Learning & development</SubSectionHeader>
-            <CareersParagraph>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc vitae ornare lectus. Aenean sit amet condimentum leo. Integer sit amet odio lectus.</CareersParagraph>
-          </Brick>
-          <Brick
-            icon={healthCoverIcon}
-          >
-            <SubSectionHeader>Health cover</SubSectionHeader>
-            <CareersParagraph>We provide health cover </CareersParagraph>
-          </Brick>
-          <Brick
-            icon={parentalLeaveIcon}
-          >
-            <SubSectionHeader>Parental leave</SubSectionHeader>
-            <CareersParagraph>Vivamus nunc eros, gravida eu tempor non, finibus tempor arcu. Quisque venenatis imperdiet tellus sed varius. Praesent vitae ornare sem.</CareersParagraph>
-          </Brick>
-          <Brick
-            icon={homeOfficeIcon}
-          >
-            <SubSectionHeader>Home office & coworking</SubSectionHeader>
-            <CareersParagraph>In convallis consequat sapien, dignissim aliquet dui auctor ut. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos.</CareersParagraph>
-          </Brick>
-          <Brick
-            icon={wellnessIcon}
-          >
-            <SubSectionHeader>Wellness</SubSectionHeader>
-            <CareersParagraph>Fusce ut eleifend nibh. Cras orci diam, auctor id nulla nec, posuere pretium massa. Suspendisse pellentesque felis sit amet congue lacinia. Nunc gravida velit dolor.</CareersParagraph>
-          </Brick>
-          <Brick
-            icon={profitShareIcon}
-          >
-            <SubSectionHeader>Profit share & ownership</SubSectionHeader>
-            <CareersParagraph>Nam porttitor enim at arcu ornare facilisis. Nullam varius vehicula erat vitae egestas. In tempor ac lorem justo molestie vehicula. Proin convallis euismod risus ac feugiat.</CareersParagraph>
-          </Brick>
-        </Masonry>
+        <Benefits />
       </Subsection>
     </Section>
